Fix health repo interface  to use the models rather than proto
---Detailed changes summary----

internal/biz/health/health.go — Changed HealthRepo interface and HealthUsecase.IsBackendAvailable() to return model.HealthResult instead of *pb.GetReadyzResponse. Removed pb import, added model import.

internal/data/health/healthrepository.go — Changed IsBackendAvailable() and IsRelationsAvailable() to return model.HealthResult. Eliminated the newResponse() helper and the pb import entirely.

internal/mocks/mocks.go — Updated MockHealthRepo methods to return model.HealthResult. Removed the now-unused pb import.

internal/service/health/health.go — GetReadyz() now converts model.HealthResult to *pb.GetReadyzResponse at the service boundary, which is where protobuf conversion belongs in the clean architecture.

internal/service/health/health_test.go — Updated mock setup to use model.HealthResult instead of *pb.GetReadyzResponse.

internal/data/health/healthrepository_test.go — Updated all assertions from uint32(500)/uint32(200) to int values (500/200) matching model.HealthResult.Code.

+++ internal/data/health/healthrepository_test.go
@@ -36,16 +36,14 @@ func TestHealthInit(t *testing.T) {
 	healthRepo := New(db, relationsRepo, relationsConfig)
 	assert.NotNil(t, healthRepo)
 
-	// just check default negative case for now when using sqlite db, and empty config
-	// storage should be okay, relations api should not because tokenConfig is empty
 	resp, err := healthRepo.IsBackendAvailable(ctx)
 	assert.Nil(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Equal(t, "RELATIONS-API UNHEALTHY", resp.Status)
 
 	resp, err = healthRepo.IsRelationsAvailable(ctx)
 	assert.Nil(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Equal(t, "RELATIONS-API UNHEALTHY", resp.Status)
 }
 
@@ -60,7 +58,7 @@ func TestHealthRepo_IsBackendAvailable_AllCases(t *testing.T) {
 	assert.NotNil(t, healthRepo)
 	resp, err := healthRepo.IsBackendAvailable(ctx)
 	assert.Nil(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Equal(t, "RELATIONS-API UNHEALTHY", resp.Status)
 
 	sqlDB, _ := db.DB()
@@ -69,22 +67,20 @@ func TestHealthRepo_IsBackendAvailable_AllCases(t *testing.T) {
 	}
 	resp, err = healthRepo.IsBackendAvailable(ctx)
 	assert.Nil(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Contains(t, resp.Status, "STORAGE UNHEALTHY")
 	assert.Contains(t, resp.Status, "RELATIONS-API UNHEALTHY")
 
 	db1 := setupGorm(t)
 	simpleRelations1 := data.NewSimpleRelationsRepository()
-	// Healthy relations (default state)
 	healthRepo1 := New(db1, simpleRelations1, relationsConfig)
 	resp, err = healthRepo1.IsBackendAvailable(ctx)
 	assert.NoError(t, err)
-	assert.Equal(t, uint32(200), resp.Code)
+	assert.Equal(t, 200, resp.Code)
 	assert.Contains(t, resp.Status, "Storage type sqlite")
 
 	db2 := setupGorm(t)
 	simpleRelations2 := data.NewSimpleRelationsRepository()
-	// Healthy relations, but DB closed
 	sqlDB2, _ := db2.DB()
 	if err := sqlDB2.Close(); err != nil {
 		t.Logf("Warning: failed to close db: %v", err)
@@ -92,7 +88,7 @@ func TestHealthRepo_IsBackendAvailable_AllCases(t *testing.T) {
 	healthRepo2 := New(db2, simpleRelations2, relationsConfig)
 	resp, err = healthRepo2.IsBackendAvailable(ctx)
 	assert.NoError(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Contains(t, resp.Status, "STORAGE UNHEALTHY: sqlite")
 	assert.NotContains(t, resp.Status, "RELATIONS-API UNHEALTHY")
 
@@ -102,7 +98,7 @@ func TestHealthRepo_IsBackendAvailable_AllCases(t *testing.T) {
 	healthRepo3 := New(db3, simpleRelations3, relationsConfig)
 	resp, err = healthRepo3.IsBackendAvailable(ctx)
 	assert.NoError(t, err)
-	assert.Equal(t, uint32(500), resp.Code)
+	assert.Equal(t, 500, resp.Code)
 	assert.Equal(t, "RELATIONS-API UNHEALTHY", resp.Status)
 
 }
